test-docker-weaviate/requirements.txt
weaviate-client
fastapi
uvicorn[standard]
ollama
